read/coff: split into modules
